setup refactor to classreg experiment
