+++ pycaret/internal/pycaret_experiment/tabular_experiment.py
@@ -1,14 +1,11 @@
 import os
 import gc
 import logging
-import warnings
-import traceback
 import random
 import secrets
-import time
 import traceback
 import warnings
-from typing import List, Tuple, Any, Union, Optional, Dict
+from typing import List, Tuple, Dict
 from unittest.mock import patch
 from joblib.memory import Memory
 from packaging import version
@@ -21,19 +18,8 @@ import pycaret.internal.patches.yellowbrick
 import pycaret.internal.persistence
 import pycaret.internal.preprocess
 import scikitplot as skplt  # type: ignore
-import sklearn
-from boruta import BorutaPy
-from imblearn.over_sampling import SMOTE
 from pandas.io.formats.style import Styler
 from pycaret.internal.logging import create_logger
-from pycaret.internal.preprocess import (
-    TransfomerWrapper,
-    ExtractDateTimeFeatures,
-    EmbedTextFeatures,
-    RemoveMulticollinearity,
-    RemoveOutliers,
-    FixImbalancer,
-)
 from pycaret.internal.meta_estimators import get_estimator_from_meta_estimator
 from pycaret.internal.pipeline import Pipeline as InternalPipeline
 from pycaret.internal.pipeline import (
@@ -45,49 +31,15 @@ from pycaret.internal.pipeline import (
 from pycaret.internal.plots.helper import MatplotlibDefaultDPI
 from pycaret.internal.plots.yellowbrick import show_yellowbrick_plot
 from pycaret.internal.pycaret_experiment.pycaret_experiment import _PyCaretExperiment
-from pycaret.containers.models.classification import get_all_model_containers as get_classifiers
-from pycaret.containers.models.regression import get_all_model_containers as get_regressors
 from pycaret.internal.pycaret_experiment.utils import MLUsecase
 from pycaret.internal.utils import (
     check_features_exist,
-    get_columns_to_stratify_by,
     get_model_name,
     mlflow_remove_bad_chars,
-    normalize_custom_transformers,
-    df_shrink_dtypes,
 )
 from pycaret.internal.validation import *
 
-from category_encoders.one_hot import OneHotEncoder
-from category_encoders.ordinal import OrdinalEncoder
-from category_encoders.leave_one_out import LeaveOneOutEncoder
-from sklearn.decomposition import PCA, IncrementalPCA, KernelPCA
-from sklearn.experimental import enable_iterative_imputer
-from sklearn.feature_selection import (
-    VarianceThreshold,
-    SelectFromModel,
-    SequentialFeatureSelector,
-)
-from sklearn.impute import IterativeImputer, SimpleImputer
 from sklearn.model_selection import BaseCrossValidator  # type: ignore
-from sklearn.model_selection import (
-    GroupKFold,
-    KFold,
-    StratifiedKFold,
-    TimeSeriesSplit,
-    train_test_split,
-)
-from sklearn.preprocessing import (
-    LabelEncoder,
-    MaxAbsScaler,
-    MinMaxScaler,
-    PolynomialFeatures,
-    PowerTransformer,
-    QuantileTransformer,
-    RobustScaler,
-    StandardScaler,
-    KBinsDiscretizer,
-)
 
 
 warnings.filterwarnings("ignore")
@@ -116,6 +68,7 @@ class _TabularExperiment(_PyCaretExperiment):
                 "logging_param",
                 "log_plots_param",
                 "data",
+                "idx",
                 "gpu_param",
                 "_all_models",
                 "_all_models_internal",
@@ -408,116 +361,57 @@ class _TabularExperiment(_PyCaretExperiment):
             del _internal_pipeline_temp
         gc.collect()
 
-    def setup(
+    def _profile(self, profile, profile_kwargs):
+        if profile:
+            profile_kwargs = profile_kwargs or {}
+
+            if self.verbose:
+                print("Loading profile... Please Wait!")
+            try:
+                import pandas_profiling
+
+                self.report = pandas_profiling.ProfileReport(self.data, **profile_kwargs)
+            except Exception as ex:
+                print("Profiler Failed. No output to show, continue with modeling.")
+                self.logger.error(
+                    f"Data Failed with exception:\n {ex}\n"
+                    "No output to show, continue with modeling."
+                )
+
+    def _initialize_setup(
         self,
-        data: pd.DataFrame,
-        target: Union[int, str] = -1,
-        train_size: float = 0.7,
-        test_data: Optional[pd.DataFrame] = None,
-        ordinal_features: Optional[Dict[str, list]] = None,
-        numeric_features: Optional[List[str]] = None,
-        categorical_features: Optional[List[str]] = None,
-        date_features: Optional[List[str]] = None,
-        text_features: Optional[List[str]] = None,
-        ignore_features: Optional[List[str]] = None,
-        keep_features: Optional[List[str]] = None,
-        preprocess: bool = True,
-        imputation_type: str = "simple",
-        numeric_imputation: str = "mean",
-        categorical_imputation: str = "constant",
-        iterative_imputation_iters: int = 5,
-        numeric_iterative_imputer: Union[str, Any] = "lightgbm",
-        categorical_iterative_imputer: Union[str, Any] = "lightgbm",
-        text_features_method: str = "tf-idf",
-        max_encoding_ohe: int = 5,
-        encoding_method: Optional[Any] = None,
-        polynomial_features: bool = False,
-        polynomial_degree: int = 2,
-        low_variance_threshold: float = 0,
-        remove_multicollinearity: bool = False,
-        multicollinearity_threshold: float = 0.9,
-        bin_numeric_features: Optional[List[str]] = None,
-        remove_outliers: bool = False,
-        outliers_method: str = "iforest",
-        outliers_threshold: float = 0.05,
-        fix_imbalance: bool = False,
-        fix_imbalance_method: Optional[Any] = None,
-        transformation: bool = False,
-        transformation_method: str = "yeo-johnson",
-        normalize: bool = False,
-        normalize_method: str = "zscore",
-        pca: bool = False,
-        pca_method: str = "linear",
-        pca_components: Union[int, float] = 1.0,
-        feature_selection: bool = False,
-        feature_selection_method: str = "classic",
-        feature_selection_estimator: Union[str, Any] = "lightgbm",
-        n_features_to_select: int = 10,
-        transform_target=False,
-        transform_target_method="box-cox",
-        custom_pipeline: Any = None,
-        data_split_shuffle: bool = True,
-        data_split_stratify: Union[bool, List[str]] = False,  # added in pycaret==2.2
-        fold_strategy: Union[str, Any] = "kfold",  # added in pycaret==2.2
-        fold: int = 10,  # added in pycaret==2.2
-        fh: Union[List[int], int, np.array] = 1,
-        fold_shuffle: bool = False,
-        fold_groups: Optional[Union[str, pd.DataFrame]] = None,
         n_jobs: Optional[int] = -1,
-        use_gpu: bool = False,  # added in pycaret==2.1
+        use_gpu: bool = False,
         html: bool = True,
         session_id: Optional[int] = None,
         system_log: Union[bool, logging.Logger] = True,
         log_experiment: bool = False,
         experiment_name: Optional[str] = None,
-        log_plots: Union[bool, list] = False,
-        log_profile: bool = False,
-        log_data: bool = False,
-        silent: bool = False,
-        seasonal_period: Optional[int] = None,
         memory: Union[bool, str, Memory] = True,
         verbose: bool = True,
-        profile: bool = False,
-        profile_kwargs: Dict[str, Any] = None,
-        display: Optional[Display] = None,
     ):
         """
-        This function initializes the environment in pycaret and creates the
-        transformation pipeline to prepare the data for modeling and deployment.
+        This function initializes the environment in pycaret.
         setup() must called before executing any other function in pycaret. It
         takes only two mandatory parameters: data and name of the target column.
 
         """
         from pycaret.utils import __version__
 
-        # Settings ================================================= >>
-
-        pd.set_option("display.max_columns", 100)
-        pd.set_option("display.max_rows", 100)
-
-        # Attribute definition ===================================== >>
-
         # Parameter attrs
-        self.data = data
-        self.target_param = target
-        self.transform_target_param = transform_target
-        self.transform_target_method_param = transform_target_method
         self.n_jobs_param = n_jobs
         self.gpu_param = use_gpu
-        self.fold_param = fold
-        self.fold_groups_param = None
         self.html_param = html
         self.logging_param = log_experiment
-        self.log_plots_param = log_plots or False
         self.memory = get_memory(memory)
+        self.verbose = verbose
 
         # Global attrs
+        self.USI = secrets.token_hex(nbytes=2)
         self.seed = random.randint(150, 9000) if session_id is None else session_id
         np.random.seed(self.seed)
 
         # Initialization =========================================== >>
-
-        runtime_start = time.time()
 
         # Get local parameters to write to logger
         function_params_str = ", ".join(
@@ -537,711 +431,11 @@ class _TabularExperiment(_PyCaretExperiment):
         self.logger.info(f"ML Usecase: {self._ml_usecase}")
         self.logger.info(f"version {__version__}")
         self.logger.info("Initializing setup()")
-        self.logger.info(f"setup(data=data, {function_params_str})")
-
-        self.USI = secrets.token_hex(nbytes=2)
         self.logger.info(f"self.USI: {self.USI}")
 
         self.logger.info(f"self.variable_keys: {self.variable_keys}")
 
         self._check_enviroment()
-
-        # Checking parameters ====================================== >>
-
-        self.logger.info("Checking parameters...")
-
-        if not isinstance(data, pd.DataFrame):
-            raise TypeError("The provided data must be of type pandas.DataFrame.")
-        elif data.empty:
-            raise ValueError("The provided data cannot be an empty dataframe.")
-
-        if train_size <= 0 or train_size > 1:
-            raise ValueError("train_size parameter has to be positive and not above 1.")
-
-        # Checking target parameter
-        if not self._is_unsupervised():
-            if isinstance(target, str):
-                if target not in data.columns:
-                    raise ValueError(
-                        "Invalid value for the target parameter. "
-                        f"Column {target} not found in the data."
-                    )
-                self.target_param = target
-            else:
-                self.target_param = data.columns[target]
-
-        # checking session_id
-        if session_id is not None:
-            if type(session_id) is not int:
-                raise TypeError("session_id parameter must be an integer.")
-
-        if profile_kwargs is None:
-            profile_kwargs = {}
-        elif not isinstance(profile_kwargs, dict):
-            raise TypeError("profile_kwargs can only be a dict.")
-
-        # stratify
-        if data_split_stratify:
-            if (
-                type(data_split_stratify) is not list
-                and type(data_split_stratify) is not bool
-            ):
-                raise TypeError(
-                    "data_split_stratify parameter only accepts a bool or a list of strings."
-                )
-
-            if not data_split_shuffle:
-                raise TypeError(
-                    "data_split_stratify parameter requires data_split_shuffle to be set to True."
-                )
-
-        possible_fold_strategy = [
-            "kfold",
-            "stratifiedkfold",
-            "groupkfold",
-            "timeseries",
-        ]
-        possible_time_series_fold_strategies = ["expanding", "sliding", "rolling"]
-
-        if self._ml_usecase != MLUsecase.TIME_SERIES:
-            if not (
-                fold_strategy in possible_fold_strategy
-                or is_sklearn_cv_generator(fold_strategy)
-            ):
-                raise TypeError(
-                    f"fold_strategy parameter must be either a scikit-learn compatible CV generator object or one of {', '.join(possible_fold_strategy)}."
-                )
-        elif self._ml_usecase == MLUsecase.TIME_SERIES:
-            if not (
-                fold_strategy in possible_time_series_fold_strategies
-                or is_sklearn_cv_generator(fold_strategy)
-            ):
-                raise TypeError(
-                    f"fold_strategy parameter must be either a sktime compatible CV generator object or one of '{', '.join(possible_time_series_fold_strategies)}'."
-                )
-
-        if fold_strategy == "groupkfold" and (
-            fold_groups is None or len(fold_groups) == 0
-        ):
-            raise ValueError(
-                "'groupkfold' fold_strategy requires 'fold_groups' to be a non-empty array-like object."
-            )
-
-        if isinstance(fold_groups, str):
-            if fold_groups not in self.X.columns:
-                raise ValueError(
-                    f"Column {fold_groups} used for fold_groups is not present in the dataset."
-                )
-
-        # checking fold parameter
-        if not isinstance(fold, int):
-            raise TypeError("fold parameter only accepts integer value.")
-
-        # checking fh parameter
-        if not isinstance(fh, (int, list, np.ndarray)):
-            raise TypeError(
-                f"fh parameter accepts integer. list or np.array value. Provided values is {type(fh)}"
-            )
-
-        # fold_shuffle
-        if type(fold_shuffle) is not bool:
-            raise TypeError("fold_shuffle parameter only accepts True or False.")
-
-        # log_plots
-        if isinstance(log_plots, list):
-            for i in log_plots:
-                if i not in self._available_plots:
-                    raise ValueError(
-                        f"Incorrect value for log_plots '{i}'. Possible values are: {', '.join(self._available_plots.keys())}."
-                    )
-        elif type(log_plots) is not bool:
-            raise TypeError("log_plots parameter must be a bool or a list.")
-
-        # log_data
-        if type(log_data) is not bool:
-            raise TypeError("log_data parameter only accepts True or False.")
-
-        # check transform_target
-        if type(transform_target) is not bool:
-            raise TypeError("transform_target parameter only accepts True or False.")
-
-        # transform_target_method
-        allowed_transform_target_method = ["box-cox", "yeo-johnson"]
-        if transform_target_method not in allowed_transform_target_method:
-            raise ValueError(
-                f"transform_target_method parameter only accepts {', '.join(allowed_transform_target_method)}."
-            )
-
-        # Data preparation ========================================= >>
-
-        # Standardize dataframe types to save memory
-        self.data = df_shrink_dtypes(self.data)
-
-        # Features to be ignored (are not read by self.dataset, self.X, etc...)
-        self._ign_cols = ignore_features if ignore_features else []
-
-        # Ordinal features
-        if ordinal_features:
-            check_features_exist(ordinal_features.keys(), self.X)
-        else:
-            ordinal_features = {}
-
-        # Numerical features
-        if numeric_features:
-            check_features_exist(numeric_features, self.X)
-        else:
-            numeric_features = list(self.X.select_dtypes(include="number").columns)
-
-        # Date features
-        if date_features:
-            check_features_exist(date_features, self.X)
-        else:
-            date_features = list(self.X.select_dtypes(include="datetime").columns)
-
-        # Text features
-        if text_features:
-            check_features_exist(text_features, self.X)
-        else:
-            text_features = []
-
-        # Categorical features
-        if categorical_features:
-            check_features_exist(categorical_features, self.X)
-        else:
-            # Default should exclude datetime and text columns
-            categorical_features = [
-                col for col in self.X.select_dtypes(include=["object", "category"]).columns
-                if col not in date_features + text_features
-            ]
-
-        # Features to keep during all preprocessing
-        keep_features = keep_features if keep_features else []
-
-        if self._ml_usecase == MLUsecase.TIME_SERIES:
-            pass #self.data, self.idx = self._get_data_and_idx()  # TODO: Make method
-        else:
-            self.data, self.idx = self._get_data_and_idx(
-                train_size=train_size,
-                test_data=test_data,
-                shuffle=data_split_shuffle,
-                stratify=data_split_stratify,
-            )
-
-        # Preprocessing ============================================ >>
-
-        # Initialize empty pipeline
-        self._internal_pipeline = InternalPipeline(
-            steps=[("placeholder", None)],
-            memory=self.memory,
-        )
-
-        if preprocess:
-
-            self.logger.info("Preparing preprocessing pipeline...")
-
-            # Encode target variable =============================== >>
-
-            if self.y.dtype.kind not in "ifu":
-                self._internal_pipeline.steps.append(("label_encoder", LabelEncoder()))
-
-            # Date feature engineering ============================= >>
-
-            # TODO: Could be improved allowing the user to choose which features to add
-            if date_features:
-                self.logger.info("Extracting features from datetime columns")
-                date_estimator = TransfomerWrapper(
-                    transformer=ExtractDateTimeFeatures(),
-                    include=date_features,
-                )
-
-                self._internal_pipeline.steps.append(
-                    ("date_feature_extractor", date_estimator),
-                )
-
-            # Imputation =========================================== >>
-
-            if self.data.isna().any().any():
-                # Checking parameters
-                num_dict = {"zero": "constant", "mean": "mean", "median": "median"}
-                if numeric_imputation not in num_dict:
-                    raise ValueError(
-                        "Invalid value for the numeric_imputation parameter, "
-                        f"got {numeric_imputation}. Possible values are "
-                        f"{' '.join(num_dict)}."
-                    )
-
-                cat_dict = {"constant": "constant", "mode": "most_frequent"}
-                if categorical_imputation not in cat_dict:
-                    raise ValueError(
-                        "Invalid value for the categorical_imputation "
-                        f"parameter, got {categorical_imputation}. Possible "
-                        f"values are {' '.join(cat_dict)}."
-                    )
-
-                if imputation_type == "simple":
-                    self.logger.info("Setting up simple imputation")
-
-                    num_estimator = TransfomerWrapper(
-                        transformer=SimpleImputer(
-                            strategy=num_dict[numeric_imputation],
-                            fill_value=0,
-                        ),
-                        include=numeric_features,
-                    )
-                    cat_estimator = TransfomerWrapper(
-                        transformer=SimpleImputer(
-                            strategy=cat_dict[categorical_imputation],
-                            fill_value="not_available",
-                        ),
-                        include=categorical_features,
-                    )
-
-                elif imputation_type == "iterative":
-                    self.logger.info("Setting up iterative imputation")
-
-                    # TODO: Fix iterative imputer for categorical columns
-
-                    # Dict of all regressor models available
-                    regressors = {k: v for k, v in get_regressors(self).items() if not v.is_special}
-
-                    if isinstance(numeric_iterative_imputer, str):
-                        if numeric_iterative_imputer not in regressors:
-                            raise ValueError(
-                                "Invalid value for the numeric_iterative_imputer "
-                                f"parameter, got {numeric_iterative_imputer}. "
-                                f"Allowed estimators are: {', '.join(regressors)}."
-                            )
-                        numeric_iterative_imputer = regressors[numeric_iterative_imputer].class_def()
-                    elif not hasattr(numeric_iterative_imputer, "predict"):
-                        raise ValueError(
-                            "Invalid value for the numeric_iterative_imputer "
-                            "parameter. The provided estimator does not adhere "
-                            "to sklearn's API."
-                        )
-
-                    if isinstance(categorical_iterative_imputer, str):
-                        if categorical_iterative_imputer not in regressors:
-                            raise ValueError(
-                                "Invalid value for the categorical_iterative_imputer "
-                                "parameter, got {categorical_iterative_imputer}. "
-                                f"Allowed estimators are: {', '.join(regressors)}."
-                            )
-                        categorical_iterative_imputer = regressors[categorical_iterative_imputer].class_def()
-                    elif not hasattr(categorical_iterative_imputer, "predict"):
-                        raise ValueError(
-                            "Invalid value for the categorical_iterative_imputer "
-                            "parameter. The provided estimator does not adhere "
-                            "to sklearn's API."
-                        )
-
-                    num_estimator = TransfomerWrapper(
-                        transformer=IterativeImputer(
-                            estimator=numeric_iterative_imputer,
-                            max_iter=iterative_imputation_iters,
-                            random_state=self.seed,
-                        ),
-                        include=numeric_features,
-                    )
-                    cat_estimator = TransfomerWrapper(
-                        transformer=IterativeImputer(
-                            estimator=categorical_iterative_imputer,
-                            max_iter=iterative_imputation_iters,
-                            initial_strategy="most_frequent",
-                            random_state=self.seed,
-                        ),
-                        include=categorical_features,
-                    )
-                else:
-                    raise ValueError(
-                        "Invalid value for the imputation_type parameter, got "
-                        f"{imputation_type}. Possible values are: simple, iterative."
-                    )
-
-                self._internal_pipeline.steps.extend(
-                    [
-                        ("numerical_imputer", num_estimator),
-                        ("categorical_imputer", cat_estimator),
-                    ],
-                )
-
-            # Text embedding ======================================= >>
-
-            if text_features:
-                self.logger.info("Setting text embedding...")
-                if text_features_method.lower() in ("bow", "tfidf", "tf-idf"):
-                    embed_estimator = TransfomerWrapper(
-                        transformer=EmbedTextFeatures(method=text_features_method),
-                        include=text_features,
-                    )
-                else:
-                    raise ValueError(
-                        "Invalid value for the text_features_method "
-                        "parameter. Choose between bow (Bag of Words) "
-                        f"or tf-idf, got {text_features_method}."
-                    )
-
-                self._internal_pipeline.steps.append(
-                    ("text_embedding", embed_estimator)
-                )
-
-            # Encoding ============================================= >>
-
-            self.logger.info("Setting up encoding")
-
-            # Select columns for different encoding types
-            one_hot_cols, rest_cols = [], []
-            for col in categorical_features:
-                n_unique = self.X[col].nunique()
-                if n_unique == 2:
-                    ordinal_features[col] = list(self.X[col].dropna().unique())
-                elif n_unique <= max_encoding_ohe:
-                    one_hot_cols.append(col)
-                else:
-                    rest_cols.append(col)
-
-            if ordinal_features:
-                self.logger.info("Setting up encoding of ordinal features")
-
-                # Check provided features and levels are correct
-                mapping = {}
-                for key, value in ordinal_features.items():
-                    if self.X[key].nunique() != len(value):
-                        raise ValueError(
-                            "The levels passed to the ordinal_features parameter "
-                            "doesn't match with the levels in the dataset."
-                        )
-                    for elem in value:
-                        if elem not in self.X[key].unique():
-                            raise ValueError(
-                                f"Feature {key} doesn't contain the {elem} element."
-                            )
-                    mapping[key] = {v: i for i, v in enumerate(value)}
-
-                    # Encoder always needs mapping of NaN value
-                    if np.NaN not in mapping[key]:
-                        mapping[key][np.NaN] = -1
-
-                ord_estimator = TransfomerWrapper(
-                    transformer=OrdinalEncoder(
-                        mapping=[{"col": k, "mapping": val} for k, val in mapping.items()],
-                        handle_missing="return_nan",
-                        handle_unknown="value",
-                    ),
-                    include=list(ordinal_features.keys()),
-                )
-
-                self._internal_pipeline.steps.append(
-                    ("ordinal_encoding", ord_estimator)
-                )
-
-            if categorical_features:
-                self.logger.info("Setting up encoding of categorical features")
-
-                if len(one_hot_cols) > 0:
-                    onehot_estimator = TransfomerWrapper(
-                        transformer=OneHotEncoder(
-                            use_cat_names=True,
-                            handle_missing="return_nan",
-                            handle_unknown="value",
-                        ),
-                        include=one_hot_cols,
-                    )
-
-                    self._internal_pipeline.steps.append(
-                        ("onehot_encoding", onehot_estimator)
-                    )
-
-                # Encode the rest of the categorical columns
-                if len(rest_cols) > 0:
-                    if not encoding_method:
-                        encoding_method = LeaveOneOutEncoder(
-                            handle_missing="return_nan",
-                            handle_unknown="value",
-                            random_state=self.seed,
-                        )
-
-                    rest_estimator = TransfomerWrapper(
-                        transformer=encoding_method,
-                        include=rest_cols,
-                    )
-
-                    self._internal_pipeline.steps.append(
-                        ("rest_encoding", rest_estimator)
-                    )
-
-            # Polynomial features ================================== >>
-
-            if polynomial_features:
-                self.logger.info("Setting up polynomial features")
-                polynomial = TransfomerWrapper(
-                    transformer=PolynomialFeatures(
-                        degree=polynomial_degree,
-                        interaction_only=False,
-                        include_bias=False,
-                        order="C",
-                    ),
-                )
-
-                self._internal_pipeline.steps.append(
-                    ("polynomial_features", polynomial)
-                )
-
-            # Low variance ========================================= >>
-
-            if low_variance_threshold:
-                self.logger.info("Setting up variance threshold")
-                if low_variance_threshold < 0:
-                    raise ValueError(
-                        "Invalid value for the ignore_low_variance parameter. "
-                        f"The value should be >0, got {low_variance_threshold}."
-                    )
-                else:
-                    variance_estimator = TransfomerWrapper(
-                        transformer=VarianceThreshold(low_variance_threshold),
-                        exclude=keep_features,
-                    )
-
-                self._internal_pipeline.steps.append(
-                    ("low_variance", variance_estimator)
-                )
-
-            # Remove multicollinearity ============================= >>
-
-            if remove_multicollinearity:
-                self.logger.info("Setting up removing multicollinearity")
-                if 0 > multicollinearity_threshold or multicollinearity_threshold > 1:
-                    raise ValueError(
-                        "Invalid value for the multicollinearity_threshold "
-                        "parameter. Value should lie between 0 and 1, got "
-                        f"{multicollinearity_threshold}."
-                    )
-
-                multicollinearity = TransfomerWrapper(
-                    transformer=RemoveMulticollinearity(multicollinearity_threshold),
-                    exclude=keep_features,
-                )
-
-                self._internal_pipeline.steps.append(
-                    ("remove_multicollinearity", multicollinearity)
-                )
-
-            # Bin numerical features =============================== >>
-
-            if bin_numeric_features:
-                self.logger.info("Setting up binning of numerical features")
-                check_features_exist(bin_numeric_features, self.X)
-
-                binning_estimator = TransfomerWrapper(
-                    transformer=KBinsDiscretizer(encode="ordinal", strategy="kmeans"),
-                    include=bin_numeric_features,
-                )
-
-                self._internal_pipeline.steps.append(
-                    ("bin_numeric_features", binning_estimator)
-                )
-
-            # Remove outliers ====================================== >>
-
-            if remove_outliers:
-                self.logger.info("Setting up removing outliers")
-                if outliers_method.lower() not in ("iforest", "ee", "lof"):
-                    raise ValueError(
-                        "Invalid value for the outliers_method parameter, "
-                        f"got {outliers_method}. Possible values are: "
-                        "'iforest', 'ee' or 'lof'."
-                    )
-
-                outliers = TransfomerWrapper(
-                    RemoveOutliers(
-                        method=outliers_method,
-                        threshold=outliers_threshold,
-                    ),
-                )
-
-                self._internal_pipeline.steps.append(
-                    ("remove_outliers", outliers)
-                )
-
-            # Balance the dataset ================================== >>
-
-            if fix_imbalance:
-                self.logger.info("Setting up imbalanced handling")
-                if fix_imbalance_method is None:
-                    balance_estimator = FixImbalancer(SMOTE())
-                elif not hasattr(fix_imbalance_method, "fit_resample"):
-                    raise ValueError(
-                        "Invalid value for the fix_imbalance_method parameter. "
-                        "The provided value must be a imblearn estimator, got "
-                        f"{fix_imbalance_method.__class__.__name_}."
-                    )
-                else:
-                    balance_estimator = FixImbalancer(fix_imbalance_method)
-
-                balance_estimator = TransfomerWrapper(balance_estimator)
-                self._internal_pipeline.steps.append(("balance", balance_estimator))
-
-            # Transformation ======================================= >>
-
-            if transformation:
-                self.logger.info("Setting up column transformation")
-                if transformation_method == "yeo-johnson":
-                    transformation_estimator = PowerTransformer(
-                        method="yeo-johnson", standardize=False, copy=True
-                    )
-                elif transformation_method == "quantile":
-                    transformation_estimator = QuantileTransformer(
-                        random_state=self.seed,
-                        output_distribution="normal",
-                    )
-                else:
-                    raise ValueError(
-                        "Invalid value for the transformation_method parameter. "
-                        "The value should be either yeo-johnson or quantile, "
-                        f"got {transformation_method}."
-                    )
-
-                self._internal_pipeline.steps.append(
-                    ("transformation", TransfomerWrapper(transformation_estimator))
-                )
-
-            # Normalization ======================================== >>
-
-            if normalize:
-                self.logger.info("Setting up feature normalization")
-                norm_dict = {
-                    "zscore": StandardScaler(),
-                    "minmax": MinMaxScaler(),
-                    "maxabs": MaxAbsScaler(),
-                    "robust": RobustScaler(),
-                }
-                if normalize_method in norm_dict:
-                    normalize_estimator = TransfomerWrapper(norm_dict[normalize_method])
-                else:
-                    raise ValueError(
-                        "Invalid value for the normalize_method parameter, got "
-                        f"{normalize_method}. Possible values are: {' '.join(norm_dict)}."
-                    )
-
-                self._internal_pipeline.steps.append(("normalize", normalize_estimator))
-
-            # PCA ================================================== >>
-
-            if pca:
-                self.logger.info("Setting up PCA")
-                if pca_components <= 0:
-                    raise ValueError(
-                        "Invalid value for the pca_components parameter. "
-                        f"The value should be >0, got {pca_components}."
-                    )
-                elif pca_components <= 1:
-                    pca_components = int(pca_components * self.X.shape[1])
-                elif pca_components <= self.X.shape[1]:
-                    pca_components = int(pca_components)
-                else:
-                    raise ValueError(
-                        "Invalid value for the pca_components parameter. "
-                        "The value should be smaller than the number of "
-                        f"features, got {pca_components}."
-                    )
-
-                pca_dict = {
-                    "linear": PCA(n_components=pca_components),
-                    "kernel": KernelPCA(n_components=pca_components, kernel="rbf"),
-                    "incremental": IncrementalPCA(n_components=pca_components),
-                }
-                if pca_method in pca_dict:
-                    pca_estimator = TransfomerWrapper(
-                        transformer=pca_dict[pca_method],
-                        exclude=keep_features,
-                    )
-                else:
-                    raise ValueError(
-                        "Invalid value for the pca_method parameter, got "
-                        f"{pca_method}. Possible values are: {' '.join(pca_dict)}."
-                    )
-
-                self._internal_pipeline.steps.append(("pca", pca_estimator))
-
-            # Feature selection ==================================== >>
-
-            if feature_selection:
-                self.logger.info("Setting up feature selection...")
-
-                if self._ml_usecase == MLUsecase.CLASSIFICATION:
-                    func = get_classifiers
-                else:
-                    func = get_regressors
-
-                models = {k: v for k, v in func(self).items() if not v.is_special}
-                if isinstance(feature_selection_estimator, str):
-                    if feature_selection_estimator not in models:
-                        raise ValueError(
-                            "Invalid value for the feature_selection_estimator "
-                            f"parameter, got {feature_selection_estimator}. Allowed "
-                            f"estimators are: {', '.join(models)}."
-                        )
-                    fs_estimator = models[feature_selection_estimator].class_def()
-                elif not hasattr(feature_selection_estimator, "predict"):
-                    raise ValueError(
-                        "Invalid value for the feature_selection_estimator parameter. "
-                        "The provided estimator does not adhere to sklearn's API."
-                    )
-
-                if feature_selection_method.lower() == "classic":
-                    feature_selector = TransfomerWrapper(
-                        transformer=SelectFromModel(
-                            estimator=fs_estimator,
-                            threshold=-np.inf,
-                            max_features=n_features_to_select,
-                        ),
-                        exclude=keep_features,
-                    )
-                elif feature_selection_method.lower() == "sequential":
-                    feature_selector = TransfomerWrapper(
-                        transformer=SequentialFeatureSelector(
-                            estimator=fs_estimator,
-                            n_features_to_select=n_features_to_select,
-                            n_jobs=self.n_jobs_param,
-                        ),
-                        exclude=keep_features,
-                    )
-                elif feature_selection_method.lower() == "boruta":
-                    # TODO: Fix
-                    feature_selector = TransfomerWrapper(
-                        transformer=BorutaPy(
-                            estimator=fs_estimator,
-                            n_estimators="auto",
-                        ),
-                        exclude=keep_features,
-                    )
-                else:
-                    raise ValueError(
-                        "Invalid value for the feature_selection_method parameter, "
-                        f"got {feature_selection_method}. Possible values are: "
-                        "'classic' or 'boruta'."
-                    )
-
-                self._internal_pipeline.steps.append(
-                    ("feature_selection", feature_selector)
-                )
-
-        # Custom transformers ====================================== >>
-
-        if custom_pipeline:
-            self.logger.info("Setting up custom pipeline")
-            for name, estimator in normalize_custom_transformers(custom_pipeline):
-                self._internal_pipeline.steps.append(
-                    (name, TransfomerWrapper(estimator))
-                )
-
-        # Remove placeholder step
-        if len(self._internal_pipeline) > 1:
-            self._internal_pipeline.steps.pop(0)
-
-        self._internal_pipeline.fit(self.X_train, self.y_train)
-
-        self.logger.info(f"Finished creating preprocessing pipeline.")
-        self.logger.info(f"Pipeline: {self._internal_pipeline}")
 
         # Set up GPU usage ========================================= >>
 
@@ -1274,214 +468,6 @@ class _TabularExperiment(_PyCaretExperiment):
                     raise ImportError(message)
                 else:
                     self.logger.warning(message)
-
-        # Set up folding strategy ================================== >>
-
-        self.logger.info("Set up folding strategy.")
-
-        if fold_groups is not None:
-            if isinstance(fold_groups, str):
-                self.fold_groups_param = self.X[fold_groups]
-            else:
-                self.fold_groups_param = fold_groups
-            if pd.isnull(self.fold_groups_param).any():
-                raise ValueError(f"fold_groups cannot contain NaNs.")
-        self.fold_shuffle_param = fold_shuffle
-
-        if not self._is_unsupervised():
-            fold_random_state = self.seed if self.fold_shuffle_param else None
-
-            if self._ml_usecase != MLUsecase.TIME_SERIES:
-                if fold_strategy == "kfold":
-                    self.fold_generator = KFold(
-                        self.fold_param,
-                        random_state=fold_random_state,
-                        shuffle=self.fold_shuffle_param,
-                    )
-                elif fold_strategy == "stratifiedkfold":
-                    self.fold_generator = StratifiedKFold(
-                        self.fold_param,
-                        random_state=fold_random_state,
-                        shuffle=self.fold_shuffle_param,
-                    )
-                elif fold_strategy == "groupkfold":
-                    self.fold_generator = GroupKFold(self.fold_param)
-                elif fold_strategy == "timeseries":
-                    self.fold_generator = TimeSeriesSplit(self.fold_param)
-                else:
-                    self.fold_generator = fold_strategy
-
-            elif self._ml_usecase == MLUsecase.TIME_SERIES:
-                # Set splitter
-                self.fold_generator = None
-                if fold_strategy in possible_time_series_fold_strategies:
-                    self.fold_strategy = fold_strategy  # save for use in methods later
-                    self.fold_generator = self.get_fold_generator(fold=self.fold_param)
-                else:
-                    self.fold_generator = fold_strategy
-
-                    # Number of folds
-                    self.fold_param = fold_strategy.get_n_splits(y=self.y_train)
-
-        if self._ml_usecase == MLUsecase.TIME_SERIES:
-            from pycaret.internal.tests.time_series import (
-                recommend_uppercase_d,
-                recommend_lowercase_d,
-            )
-
-            self.white_noise = None
-            wn_results = self.check_stats(test="white_noise")
-            wn_values = wn_results.query("Property == 'White Noise'")["Value"]
-
-            # There can be multiple lags values tested.
-            # Checking the percent of lag values that indicate white noise
-            percent_white_noise = sum(wn_values) / len(wn_values)
-            if percent_white_noise == 0:
-                self.white_noise = "No"
-            elif percent_white_noise == 1.00:
-                self.white_noise = "Yes"
-            else:
-                self.white_noise = "Maybe"
-
-            self.lowercase_d = recommend_lowercase_d(data=self.y)
-            # TODO: Should sp this overrise the self.seasonal_period since sp
-            # will be used for all models and the same checks will need to be
-            # done there as well
-            sp = self.seasonal_period if self.seasonality_present else 1
-            self.uppercase_d = (
-                recommend_uppercase_d(data=self.y, sp=sp) if sp > 1 else 0
-            )
-
-        # Final display ============================================ >>
-
-        self.logger.info("Creating final display dataframe.")
-
-        if isinstance(numeric_iterative_imputer, str):
-            num_imputer = numeric_iterative_imputer
-        else:
-            num_imputer = numeric_iterative_imputer.__class__.__name__
-
-        if isinstance(categorical_iterative_imputer, str):
-            cat_imputer = categorical_iterative_imputer
-        else:
-            cat_imputer = categorical_iterative_imputer.__class__.__name__
-
-        display_container = self._get_setup_display(
-            target_type="Multiclass" if self._is_multiclass() else "Binary",
-            train_size=train_size,
-            ordinal_features=len(ordinal_features) if ordinal_features else 0,
-            numerical_features=len(numeric_features),
-            categorical_features=len(categorical_features),
-            date_features=len(date_features),
-            text_features=len(text_features),
-            ignore_features=len(self._ign_cols),
-            keep_features=len(keep_features),
-            missing_values=self.data.isna().sum().sum(),
-            preprocess=preprocess,
-            imputation_type=imputation_type,
-            numeric_imputation=numeric_imputation,
-            categorical_imputation=categorical_imputation,
-            iterative_imputation_iters=iterative_imputation_iters,
-            numeric_iterative_imputer=num_imputer,
-            categorical_iterative_imputer=cat_imputer,
-            text_features_method=text_features_method,
-            max_encoding_ohe=max_encoding_ohe,
-            encoding_method=encoding_method,
-            transformation=transformation,
-            transformation_method=transformation_method,
-            normalize=normalize,
-            normalize_method=normalize_method,
-            low_variance_threshold=low_variance_threshold,
-            remove_multicollinearity=remove_multicollinearity,
-            multicollinearity_threshold=multicollinearity_threshold,
-            remove_outliers=remove_outliers,
-            outliers_threshold=outliers_threshold,
-            polynomial_features=polynomial_features,
-            polynomial_degree=polynomial_degree,
-            fix_imbalance=fix_imbalance,
-            fix_imbalance_method=fix_imbalance_method,
-            pca=pca,
-            pca_method=pca_method,
-            pca_components=pca_components,
-            feature_selection=feature_selection,
-            feature_selection_method=feature_selection_method,
-            feature_selection_estimator=feature_selection_estimator,
-            n_features_to_select=n_features_to_select,
-            custom_pipeline=custom_pipeline,
-        )
-
-        # Other attrs
-        self.experiment__ = []
-        self.master_model_container = []
-        self._all_models, self._all_models_internal = self._get_models()
-        self._all_metrics = self._get_metrics()
-
-        if verbose:
-            print(display_container)
-
-        # Pandas profiling ========================================= >>
-
-        if profile:
-            if verbose:
-                print("Setup Successfully Completed! Loading profile... Please Wait!")
-            try:
-                import pandas_profiling
-
-                pf = pandas_profiling.ProfileReport(self.data, **profile_kwargs)
-                display.display(pf, clear=True)
-            except Exception as ex:
-                print("Profiler Failed. No output to show, continue with modeling.")
-                self.logger.error(
-                    f"Data Failed with exception:\n {ex}\n"
-                    "No output to show, continue with modeling."
-                )
-
-        # MLflow and wrap-up ======================================= >>
-
-        self._set_up_mlflow(
-            display_container,
-            np.array(time.time() - runtime_start).round(2),
-            log_profile,
-            profile_kwargs,
-            log_data,
-            display,
-        )
-
-        self.logger.info(f"self.display_container: {len(self.display_container)}")
-        self.logger.info(f"Pipeline: {str(self._internal_pipeline)}")
-        self.logger.info("setup() successfully completed............................")
-
-        # Reset pandas option
-        pd.reset_option("display.max_rows")
-        pd.reset_option("display.max_columns")
-
-        self._setup_ran = True
-
-        ## Disabling of certain metrics.
-        ## NOTE: This must be run after _setup_ran has been set, else metrics can
-        # not be retrieved.
-        if (
-            self._ml_usecase == MLUsecase.TIME_SERIES
-            and len(self.fh) == 1
-            and "r2" in self._get_metrics()
-        ):
-            # disable R2 metric if it exists in the metrics since R2 needs
-            # at least 2 values
-            self.remove_metric("R2")
-
-        self.logger.info(
-            f"self.master_model_container: {len(self.master_model_container)}"
-        )
-        self.logger.info(f"self.display_container: {len(self.display_container)}")
-
-        self.logger.info(str(self._internal_pipeline))
-        self.logger.info(
-            "setup() successfully completed......................................"
-        )
-
-        gc.collect()
-
-        return self
 
     def plot_model(
         self,
