+++ pycaret/regression.py
@@ -1,17 +1,16 @@
 # Module: Regression
 # Author: Moez Ali <moez.ali@queensu.ca>
 # License: MIT
-# Release: PyCaret 2.2.0
-# Last modified : 25/10/2020
+
 
 import logging
 import pandas as pd
-import numpy as np
+from joblib.memory import Memory
 
 from pycaret.internal.pycaret_experiment import RegressionExperiment
 from pycaret.internal.utils import check_if_global_is_not_none
 
-from typing import List, Tuple, Any, Union, Optional, Dict
+from typing import List, Any, Union, Optional, Dict
 import warnings
 
 warnings.filterwarnings("ignore")
@@ -87,8 +86,8 @@ def setup(
     log_plots: Union[bool, list] = False,
     log_profile: bool = False,
     log_data: bool = False,
-    silent: bool = False,
     verbose: bool = True,
+    memory: Union[bool, str, Memory] = True,
     profile: bool = False,
     profile_kwargs: Dict[str, Any] = None,
 ):
@@ -475,14 +474,15 @@ def setup(
         Ignored when ``log_experiment`` is not True.
 
 
-    silent: bool, default = False
-        Controls the confirmation input of data types when ``setup`` is executed. When
-        executing in completely automated mode or on a remote kernel, this must be True.
-
-
     verbose: bool, default = True
         When set to False, Information grid is not printed.
 
+
+    memory: str, bool or Memory, default=True
+        Used to cache the fitted transformers of the pipeline.
+            If False: No caching is performed.
+            If True: A default temp directory is used.
+            If str: Path to the caching directory.
 
     profile: bool, default = False
         When set to True, an interactive EDA report is displayed. 
@@ -531,6 +531,8 @@ def setup(
         remove_outliers=remove_outliers,
         outliers_method=outliers_method,
         outliers_threshold=outliers_threshold,
+        fix_imbalance=False,  # Always False for regression
+        fix_imbalance_method=None,
         transformation=transformation,
         transformation_method=transformation_method,
         normalize=normalize,
@@ -561,8 +563,8 @@ def setup(
         log_plots=log_plots,
         log_profile=log_profile,
         log_data=log_data,
-        silent=silent,
         verbose=verbose,
+        memory=memory,
         profile=profile,
         profile_kwargs=profile_kwargs,
     )
